# docs/openapi.yaml
openapi: 3.0.3
info:
  title: sandboxd /v1 API
  version: "0.3.0"
  description: >
    The public, tenant-scoped API for sandboxd. This is the only surface a
    console or external integration should use — the internal /sandbox and
    /wake endpoints are not part of this contract. Scoped by the API token
    (Bearer); when auth is disabled (the single-user default) the whole
    surface is reachable without a token.
servers:
  - url: http://127.0.0.1:9090
    description: Default loopback bind (docker-compose maps SANDBOXD_API_BIND here)
security:
  - bearerAuth: []
  - {} # auth disabled by default (single-user)
...
paths:
...
  /v1/sandboxes/{id}/files:
    parameters: [ { $ref: "#/components/parameters/Id" } ]
    get:
      operationId: listSandboxFiles
      tags: [sandboxes]
      summary: List workspace files
      parameters:
        - { in: query, name: path, schema: { type: string }, description: "Directory relative to the app dir (default '.')." }
        - { in: query, name: recursive, schema: { type: boolean } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  path: { type: string }
                  recursive: { type: boolean }
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        path: { type: string, description: "Relative to the app dir." }
                        type: { type: string, enum: [file, dir] }
                        size: { type: integer }
    put:
      operationId: putSandboxFile
      tags: [sandboxes]
      summary: Write a workspace file (atomic)
      description: >
        Writes one file into the sandbox workspace, relative to the APP dir
        (`workspace/app`), matching GET /files. Parent directories are created.
        The request body is the RAW file content (opaque bytes), not JSON.
        Reserved subtrees (`.runtimed/`, `lost+found/`) are refused and a
        symlinked leaf is never overwritten. Atomic: tmp file + rename.
      parameters:
        - in: query
          name: path
          required: true
          schema: { type: string }
          description: >
            File path relative to the app dir (e.g. `src/main.py`). Absolute
            paths and `..` are rejected.
      requestBody:
        required: true
        description: Raw file content (opaque bytes) — not JSON.
        content:
          text/plain:
            schema: { type: string }
      responses:
        "200": { description: Written }
        "400": { description: Invalid path (absolute, traversal, or reserved subtree) }
        "413": { description: File exceeds the size limit }
    delete:
      operationId: deleteSandboxFile
      tags: [sandboxes]
      summary: Delete a file or directory
      description: >
        Deletes the file (or directory, recursively) at `path`, relative to the
        app dir. The workspace root itself cannot be deleted; paths that
        traverse a symlink are refused (a symlinked leaf is unlinked, never its
        target).
      parameters:
        - in: query
          name: path
          required: true
          schema: { type: string }
          description: File or directory path relative to the app dir.
      responses:
        "200": { description: Deleted }
        "400": { description: Workspace root or symlink traversal }
        "404": { description: No such file }
    patch:
      operationId: renameSandboxFile
      tags: [sandboxes]
      summary: Rename a file or directory (in place)
      description: >
        Renames the file or directory at `path` to `name`, within its parent
        directory (no move across folders). Fails if the target name already
        exists. Symlinked paths are refused.
      parameters:
        - in: query
          name: path
          required: true
          schema: { type: string }
          description: Existing file or directory path relative to the app dir.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: New base name — no slashes, no `.`/`..`.
      responses:
        "200": { description: Renamed }
        "400": { description: Invalid name or path }
        "404": { description: No such file }
        "409": { description: Target name already exists }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
...